refactored GraqlVisitor to be called GraqlConstructor and extracted analytics code into GraqlComputeConstructor
